Allow defining initializers for reference types
Nothing quite yet calls these, and I am not even sure that this is a
good idea to add to the language per se. It is very likely
that the infrastructure built here will prove useful to define custom
deallocators (e.g. for files, raw memory, ...) even if initializers
are not exactly useful per se

tl;dr Initializers may go away but the changes under the hood to
support them are likely going to stick anyway because they enable
other useful things