Start using init members in reference types
The actual implementation is hacky and does not use the right
machinery to build the function call, which will be fixed in
a subsequent patch. But the mechanism works.

The rules are:
1) only reference types can define init
2) only one init per type
3) if an init is defined, it must be called

Rule 2) is where this mechanism is probably not the right thing to do
user-experience wise, since it is not clear that having a single
constructor is really the right thing to do, short of making a large
TypeOptions{} value type and then passing that kitchen sink list
of all possible behaviors up to the init() entry point.

It may well be that defining multiple init-style entry points
in an impl {} block is the right thing to do instead of imposing
behavior at the compiler level.

Even so, I may end up removing init() constructors while keeping the
machinery that backs them, since it may be useful to support this for
custom deallocators, which are clearly unambiguous and clearly needed.