Improve type handling in binary expressions
I am not too fond of any of these rules but they do make a lot
of expressions involving integer bit-size mixing a lot easier
to write and to compile without segfaulting in LLVM

There is still some order-dependent ambiguity, e.g.
n + 1 is the type of n, but 1 + n will default to int64
without a stronger hint from somewhere else, which does not
make things very commutative.

A better solution would be to write some kind of expression
analyzer ruleset and figure out the type of the whole thing
ahead of code generation and then coerce everything