Add a hacky way to run tests that compile to a binary
These tests will produce an executable and run it and will check for
either success or failure

Right now, this conflates compiler errors with crash at runtime, but
that is relatively easy to clean up.

This feature tries to delete the files it generates but this may
not work if the test suite crashes due to a bug, so some stale
temporary files may remain.

The main thing this does is check that assert correctly
returns 0 for true and crashes for false, so that other tests
can use asserts liberally and be sure (or at least somewhat
confident) that they will get a crash when the test fails but
that this won't tear down the whole test suite