Rename VarContext to ScopeObject
The idea is that scopes can contain more than just variables,
but also types and functions. This would allow local functions
and local types to be defined.

These would still be global as far as LLVM is concerned, but
they'd be appropriately named and language-wise, only visible
from within their owner Scope