Introduce mandatory spacing for contextual keywords
_() as a parser rule allows 0 spacing, which means the parse tree of
alloc_thing() is Alloc(_thing,()) rather than
FunctionCall(alloc_thing,()) which is the much more human obvious
meaning. In general, the lack of mandatory spacing means that
funcfoo would parsed as-if declaring a function named foo and not
an identifier "funcfoo". Introduce a new rule __() which means
at least one whitespace must be present, but any number is possible

This allows the parser to do the obviously intended thing in a number
of otherwise ambiguous cases