Introduce bill of materials import mechanism
A bill of materials is a list of exported entities provided by a
C± file. In this initial commit, the bill of materials only includes
free functions and type aliases, since these are pretty easy to store
and retrieve.

This patch is sufficient to craft a BOM file for a simple program
and import somewhere else, link and run the resulting binary

It introduces the `export` attribute for symbols, and only the symbols
that are declared with the attribute will be stored in the BOM

This does not come with a test case since we don't yet have the
structure for tests that join multiple files and use the linker

The BOM is saved as JSON since this is a simple and easy to grok
file format, which would make it possible to craft all kinds
of validation and inspection for a bill of materials