Support storing variables in bill of materials
This commits adds the ability to export global variables via a bill
of materials and importing in a different unit

With this change, exchanging types via BOM is the only missing
capability. This is likely to be added progressively over the next
series of patches, starting from member values only and then adding
init, dealloc, and methods - possibly even starting from value types

Once that is done, the next step will be to rework the test system
enough to allow writing tests that export and import things via a BOM
and verifying that the resulting compiled output is coherent and works
as intended. This is mostly a test automation problem