Add driver tests
Driver tests use the new driver API to build a.out programs from one
or more files and include bom generation support. Because they are
configured via JSON, they are more flexible than either JIT or a.out
tests and can be extended in the future to support additional features
such as multiple targets, link to external libraries, more compiler
or linker flags, ...

For now, they are used to test that exporting a function and a variable
via BOM work when linking a single binary

These tests generate in-tree output, but that output is then ignored
by git, so it doesn't get checked into the repository. It may be
worthwhile to cleanup the test directory so that any stale bom
or binary do not confuse things at runtime, but that will be a future
extension