Rework most of the error-checking inline tests to use the driver API
The driver API actually allows to look for what errors are happening
which makes for more useful tests. A few remain and will be converted
in a subsequent patch, so that the only inline tests that remain are
the ones that check for proper execution. Those should be converted
to out-of-band JIT tests over time but that is less urgent