Implement refcounting in C vs. IR
This is a rewrite of the code that does inc/dec/getref in C instead
of LLVM IR. This code is much simpler to reason about and will be
much simpler to make atomic. In a.out mode, this simply involves
building one C file and linking everything else with it

This does not work as trivially in JIT mode, which means right now
there are two copies of this code lurking around. This is only a
short-term inconvenience as the plan is to get rid of the LLVM
JIT mode, and instead generate and run a temporary binary when
asked to JIT. This has numerous benefits, including that one
can JIT multiple files trivially, and that there only exists one
path to compile everything

This will be the object of subsequent patches