Replace old JIT mode with a multi-jit mode
Multi-JIT compiles an a.out binary and runs it instead of using
the LLVM JIT engine. Adapt JIT tests to use multi-JIT mode.

This change means that there is one single compilation and execution
path for all programs. Follow-up commits will delete the legacy JIT
code and simplify a lot of the compiler behavior around this