Slowly but surely migrate JIT tests to pass
This is done by a wise mix of adding assertions in some tests instead
of using return values, simplifying some tests, and outright deleting
a small number that otherwise don't add value as they are substantial
replicas of other tests

The expectation is that ultimately this will be a patch of a decent
size but with very little actual logic embedded, mostly just some
adaptation of existing code

One last vestige of the old system is that some 40-odd tests still
return int64 from main. This is of course wrong and broken, but
since the bits are all 0, it won't affect much in practice and it is
a return from main() anyway, so there's precious little left
to break anyway. It will be fixed later