Mark imported types as exported
This is done not so that types can be re-exported, but so that one
can add impls{} to an imported type.

This can occasionally be useful but it does run into the risk of
conflicting declarations in different compile units for a method
of the same name. It may be useful to add a marker to a type to
say it's "sealed", i.e. will not allow any other compile unit to
add impls{} to it