Introduce type metadata
Metadata is (for now) a pair of a type name and a deallocator. With
this patch, one word is added to every reference type that is a
pointer to the metadata. With a subsequent commit, two words will be
reclaimed for the user/sys deallocator and the metadata will just be
used to discover these during runtime

The metadata is unique per type instead of per instance, and thus can
cheaply be extended to add more information when/if needed without
overhead on each object stored