Mutating methods on value types
Value types had immutable methods (i.e. any changes to the contents
of self would be discarded on exit), but this was not an intentional
choice so much as an artifact of the grammar and the fact that
value types are not represented as pointers

This is fixed now by restricting method calls to lvalues, so things
that indeed can be represented as a (local if anything) pointer. It
would be possible to open methods to rvalues again by generating a
temporary object on "rvalue"->foo(), but this is left as a future
improvement

Once all "self" arguments are lvalues, there is some extra logic added
to generate methods with a "self-as-pointer" for value type methods,
that which is implicit for ref type methods - and to ensure correct
access to fields of value types when passed as pointers, which was
previously not done

Other arguments of value types, of course, continue to be passed by
value - which is intentional and by design - and as such are immutable