Allow initializers to be provided for value types
This falls one more gap between val and ref types, where val types
would not allow an init to be defined. This is now possible and init is
called whenever supplied, much like for reference types.

dealloc is still not possible for a value type. To some extent this is
intentional as there is no notion of reference counting for them and
no system deallocator involved in freeing memory. While some user code
could want to run at scope exit of a value type, this is likely to
introduce a lot of complexity for how copies are handled so it's worth
thinking about as a backlog item