Preliminary code to handle overloaded method calls
This is still untested and provisional but it works pretty much as
follows:
- split a call arguments into two steps:
  - constant and hintable;
  - not

The former are things like "3", "3 + 4", which can be deduced to
different types depending on the surrounding contexts and the latter
are things like "3 as int32", "alloc Foo", which have one and only one
permissible type.

When resolving an overload call, first look for the argument count
of the call. If only one overload supports that many arguments, then
use that one and consider it solved.

If multiple overloads support the effective argument count, e.g.
foo(x: int32) and foo(x: float32), then look through the constant and
hintable arguments, attempting to generate each as the type that the
function expects. If a mismatch occurs, then consider it a failed
attempt and give up on the entire candidate.

If zero or >1 candidates remain at this stage, this is an error and
diagnostics will need to be added for it

The test cases are quite minimal and only cover the happy path in
obvious cases