Add a test for an import bug
It is not possible to export/import pointers to basic types. This is
due to a combination of three bugs, two easy and one hard:
- easy, the code never goes and looks for builtin types when resolving
  an alias;
- easy, the formatting for a descriptor for an int type needs to take
  bool and byte into account and not just print int{size} for every
  case since int1 and int8 do not exist;
- hard, descriptors for *T are exported as just T, because that makes
  memory management of reference types easier to handle, but this way
  one cannot distinguish *byte from byte, hence the test case fails

This last bug is not easy to fix since a lot of logic in the compiler
assumes that one will import ref type T {} as just "T" and have the
code itself fix it up to mean *T under the hood. Changing the export
to be *T would mean that the other logic would now overwrite it to
mean **T, in turn breaking *a lot* of things, so this bug will either
entail a very minimal fix or will entail rewriting a lot of the logic
to handle reference types support in the bowels of the compiler