Fix export_ptr_type test
This is a very small and minimal fix for this test case. It fixes the
behavior of descriptor_by_llvm_type for every case except structs and
pointers-to-structs that are reference counted types, for which it
preserves the previous behavior, i.e. for

ref type Foo {}

it generates a descriptor of Name(Foo) and not Pointer(Name(Foo))
although the latter is what is really going on under the hood

For other types, *Foo is stored indeed as Pointer(Name(Foo)) and Foo
as Name(Foo), which then allows *byte to be properly exported and
imported for what it really is; export_ptr_type tests for this behavior
so it should not regress again