Get rid of undecref_on_exit
Allocas are stashed away in one magic block under the compiler's direct
control which we know has only one exit point and no allocations that
follow it. undecref_on_exit was a hack to allow deleting reference
values before calling dealloc on them, in case they were allocated on
a path that didn't initialize them, back when allocas were emitted at
the point of translation (i.e. in the same basic block where they were
defined in source code). Now that allocas are all stashed in that one
place to begin with, this failure mode can no longer exist since all
allocas of reference type are centrally initialized to null and a
decref of null is by definition boring yet safe.

tl;dr there is no need for undecref_on_exit in the current model of
variable allocation and it was one of the ugliest hacks in the codebase
so getting rid of it is a win