Remove mandatory parentheses in if/while/do statements
There is no strong need for () in these conditions other than C did it
that way back in the way. It is feasible to write them without the
parentheses, so allow it, but do require a space between the keyword
and the condition.

There is some complexity in the grammar to still allow if(thing) vs.
if thing, since that is an obvious unambiguous parse and it makes no
sense to demand if (thing) for this syntax. This does complicate the
parser a bit but only marginally, especially around not having _
between a conditional statement and its condition, since that will
need to be consumed by delimited_expression via its __ call.

Tests have been added for a variety of styles and usages and hopefully
they can catch any future regression in this area.