Automatic dereferencing of value instances
If a method is defined to take a foo: VT and a *VT is passed to it
automatically perform a dereference. This is done so that one can
transparently pass "self" into a method taking an instance value.

This becomes important with operator overloading, since this would
for example look like func eq(other: VT) and it may be useful to
write x == self, at which point this would become x->eq(self) but
self is a *, and so without dereferencing, you'd end up writing
x == *self for a value type, but not for a reference type.